Move to stats/base, and created a list of options
